subprojects/robotpy-commands-v2/docs/requirements.txt
# This file is intended for use on readthedocs
sphinx
sphinx-rtd-theme
robotpy-sphinx-plugin
-e .